Try to support monorepos
